Module impliar::enci

Modules

be_int

contains interpolator over sequence of vectors

demo_rng

fatorx

add/mult functions for factorization

i_mem

memory structure for contradiction of expected v. got

implie

struct used to contain element that can generate new elements based on schematic. element. Notation of new elements is in the form of parenthetical notation.

mat2sort

functions for <arr1> and <arr2> sort

ns

numerical seeds for types i32,f32 using std_rng

ohop

contains order-of-operations parser for parenthetical/underscored strings

parentnot

methods for parenthetical notation

pf_op

seq_encoder

an encoder for an i32 sequence

seqe

calculation of cheapest skew functions

skew

struct for error-correction

skewf32

f32 version of Skew

std_rng

pseudo-random number generating functions using rand crate