Struct impliar::enci::be_int::BEInt

pub struct BEInt {
    pub data: Array2<f32>,
    pub e_soln: Array1<f32>,
    pub r_soln: Array1<Option<f32>>,
    pub i: usize,
    pub fanalysis: HashMap<usize, usize>,
    pub ranalysis: HashMap<usize, Array1<f32>>,
    pub im: IMem,
}

binary error interpolator structure. Used to {possibly} get soln.

features:

brute-force

frequency-analysis search for substitution.

generalization of algorithm:

loop

accumulate

substitute

solve

Fields
data: Array2<f32>

x-values

e_soln: Array1<f32>

y-values

r_soln: Array1<Option<f32>>

value for each data variable (column)

i: usize

fanalysis: HashMap<usize, usize>

ranalysis: HashMap<usize, Array1<f32>>

im: IMem
Implementations

impl BEInt

pub fn initiaado(&mut self)

description

orderes de data es gracias y gorolobos es stela cumar mi yo si IMem.

pub fn error_term(&mut self, deflt: f32) -> f32

return

the error term

pub fn ds_arr2(&mut self) -> Array2<f32>

pub fn order_bfs(&mut self)

description

initial ordering by non-decreasing active size

pub fn solve_at(&mut self, i: usize, verbose: bool, solve_mode: usize) -> bool

return

outputs whether data can be solved

pub fn solve_contradiction(
&mut self,
i: usize,
stat: bool,
solve_mode: usize
) -> usize

description

solves subarr at [0..i + 1], outputs the number of remaining contradictions.

arguments

i := end index of subarr

stat := print statements?

solve_mode := 0 for substitution repr., 1 for representative repr.

return

number of contradictions in [0,i]

pub fn save_to_imem(&mut self, i: usize)

description

saves to im the rsoln and any contradictions (known|unknown vars allowed) in the range [0,i]

pub fn deduce_smap_keys_from_rsoln(&mut self, sm: HashMap<usize, Array1<f32>>)

description

determines variable solutions from cases of substitution map sm that are certain (no remaining unknown)

pub fn accumulate(&mut self, si: usize, ei: usize) -> Array1<f32>

description

accumulates the samples in the range [si,ei].

Output is [var coeff|error coeff|output].

Output is according to the running solution of known variables.

arguments

si := start index

ei := end index

return

literal value <arr1<f32>> of summation of subvector data\[si..ei + 1\]

pub fn apply_running_soln_to_expr(
&mut self,
expr: Array1<f32>,
rs: Array1<f32>
) -> Array1<f32>

arguments

expr := [var coeff|error coeff|output]

rs := [var values|error value]

return

an updated expr of unknown after running soln rs is applied

pub fn check_varsub_contradiction(
&mut self,
k: usize,
substitution_map: HashMap<usize, Array1<f32>>
) -> bool

description

checks that variable k does not contain a contradictory substitution. helper method for above method is_contradictory_substitution_map.

pub fn bfs_sub_on_var(
&mut self,
k: usize,
substitution_map: HashMap<usize, Array1<f32>>,
save_data: bool
) -> HashSet<usize>

description

performs a BFS on the substitution of variable k; outputs the hashset of all substitution variables for variable k.

pub fn substitute_solve_chain(
&mut self,
expr: Array1<f32>,
substitution_map: HashMap<usize, Array1<f32>>,
verbose: bool
) -> (Array1<Option<f32>>, i32)

description

outputs (soln, number of new variables known)

arguments

expr := [var coeff|error coeff|output]

substitution_map := variable -> substitution

verbose := print statement?

pub fn substitution_repr(
&mut self,
s: Array1<f32>,
start_ref: usize,
end_ref: usize,
verbose: bool
) -> HashMap<usize, Array1<f32>>

description

outputs a substitution representation of the sample, a vector of length |s|; substitution minimizes the number of (active && unknown) variables. Derives variable substitutes from the chunk [start_ref,end_ref]. Substitution targets variables that are not known in the running solution. Uses a brute-force approach.

argument

s := expression

start_ref := start index

end_ref := end index

verbose := print statements?

return

variable -> <arr1<f32>> (substitution)

pub fn conduct_substitution(
&mut self,
expr: Array1<f32>,
substitution_map: HashMap<usize, Array1<f32>>,
verbose: bool
) -> Array1<f32>

description

converts the expression into an output expression by the substitution map; if the substitution map does not contain a variable i, then output expression will use identity;

warning

substitution algorithm will progressively substitute each variable that is present as a key in the substitution map; if substitution map is not structured properly, may lead to infinite loop.

arguments

expr := [var coefficients|error term|output] substitution_map := variable -> substitution variables verbose := print statements?

return

<arr1<f32>> that is the substitution

pub fn conduct_substitution_(
&mut self,
expr: Array1<f32>,
substitution_map: HashMap<usize, Array1<f32>>
) -> Array1<f32>

description

helper function; conducts 1 round of substitution. substitution checks for known vars, and output deducts the sum of the known vars

pub fn solve_by_unknown_vars(
&mut self,
unknown_vec: Array1<f32>,
rs: Array1<f32>,
wanted_value: f32
) -> bool

description

plugs in values for unknown in unknown_vec so that the update running solution to rs equals wanted_value

arguments

unknown_vec := arr1 of unknown variable coefficients, length is data.dim().1

rs := arr1 of running soln values

wanted_value := target value

return

plugging in values for unknown to rs yields wanted_value?

pub fn save_sol_to_rsoln(&mut self, soln: Array1<f32>) -> bool

arguments

soln := values of variables

return

soln active vars do not collide with rsoln?

pub fn deduce_elements_from_rsoln(
&mut self,
si: usize,
ei: usize,
verbose: bool
) -> bool

description

uses known variables in rsoln to determine values of error term and other variables; outputs whether subarr2 [si..ei + 1] can be solved. If can be solved, then saves soln.

pub fn rsoln_output(&mut self, si: usize, ei: usize) -> Array1<f32>

return

literal f32 values for subarr `data[si..ei + 1]

pub fn contradictions_in_range(
&mut self,
si: usize,
ei: usize,
absolut: bool,
verbose: bool
) -> HashSet<usize>

desription

determines the indices of elements with all active && (known if absolut) variables known that contradict e_soln

pub fn var_repr(&mut self, si: usize, vi: usize) -> Option<Array1<f32>>

description

representation of variable vi by element si. outputs an expression of the form [var coefficients|error term|output] for variable vi;if vi == 0, then None.

arguments

si := sample index

vi := variable index

pub fn var_reprs_in_range(
&mut self,
vi: usize,
si: usize,
ei: usize
) -> Vec<Array1<f32>>

description

gathers all possible representations of variable vi in range.

note

no identity repr

pub fn var_reprs_in_range_filtered(
&mut self,
vi: usize,
si: usize,
ei: usize,
excluded_vars: HashSet<usize>
) -> Vec<Array1<f32>>

return

all possible representations of variable vi in range [si,ei] that do not use any of the excluded_vars.

pub fn is_contradictory_substitution_map(
&mut self,
substitution_map: HashMap<usize, Array1<f32>>
) -> bool

description

determines if substitution map is contradictory.

If there are two variables v1,v2 in the map in which either v2 substitutes for v1 or vice-versa, then that is a contradiction.

return

contradictory outputs true, o.w. false.

pub fn running_soln_of_sample(&mut self, v: Array1<f32>) -> Array1<f32>

arguments

v := length is data.dim().1 + 1

return

literal f32 values of v

pub fn active_size_of_expr(&mut self, expr: Array1<f32>) -> usize

description

calculates the active size of the var coefficients in expr.

argument

expr := [var coefficients|error term|output]

pub fn active_size_of_soln(&mut self, soln: Array1<Option<f32>>) -> usize

description

calculates the active size of the var coefficients in soln. expr := [var coefficients|error term|output]

pub fn active_indices_of_expr(&mut self, expr: Array1<f32>) -> HashSet<usize>

return

the active indices of expr

pub fn active_indices_of_soln(
&mut self,
soln: Array1<Option<f32>>
) -> HashSet<usize>

return

the active indices of the var coefficients in soln

pub fn representative_indices_of_expr(
&mut self,
expr: Array1<f32>,
sm: HashMap<usize, Array1<f32>>
) -> HashSet<usize>

pub fn representative_indices_of_smap(
&mut self,
sm: HashMap<usize, Array1<f32>>
) -> HashSet<usize>

description

representative indices of smap are those indices found in the substitution of each key in sm but are not keys themselves.

pub fn remaining_unknown_of_sample(
&mut self,
s: Array1<f32>,
r: Array1<f32>
) -> Array1<f32>

arguments

s := sample

r := running solution, length is |s| + 1.

return

array1 with non-zero unknown, length is |s|

pub fn submap_var_frequency(&mut self, sm: HashMap<usize, Array1<f32>>)

description

performs a frequency count of substitution map

return

variable -> frequency vector of other variables as substitutes

pub fn representative_table(&mut self) -> HashMap<usize, Array1<f32>>

return

column var index -> binary arr1 of length equal to data.dim().1 (1 values correspond to possible representative)

pub fn relevant_vars_of_var_in_relevance_table(
&mut self,
rt: HashMap<usize, Array1<f32>>,
i: usize
) -> Array1<f32>

description

variant of bfs_sub_on_var; obtains all variables in which var i is either a direct or indirect parent of.

pub fn relevant_vars_structure_of_var_in_relevance_table(
&mut self,
rt: HashMap<usize, Array1<f32>>,
i: usize
) -> Vec<HashSet<usize>>

description

helper method for relevant_vars_of_var_in_relevance_table

pub fn representative_relevance_table(
&mut self,
rt: HashMap<usize, Array1<f32>>,
index_order: Array1<usize>
) -> Array2<f32>

return

2-d array, row index i corresponds to variable, column index j corresponds to frequency of variable j as representative

pub fn representative_analysis_(&mut self)

description

assigns representative_table to ranalysis.

pub fn relevance_submatrix(
&mut self,
key_indices: Array1<usize>
) -> HashMap<usize, Array1<f32>>

return

frequency of representatives for each variable in key_indices

pub fn rr_submatrix(&mut self, key_indices: Array1<usize>) -> Array2<f32>

return

representative relevance submatrix

pub fn representative_analysis(
&mut self,
rrm: Array2<f32>,
index_keys: Array1<usize>
) -> Vec<(usize, f32)>

pub fn select_var_repr_by_max_candidates(
&mut self,
i: usize,
excluded_vars: HashSet<usize>
) -> Option<Array1<f32>>

description

selects the var. repr. for variable i with the lowest active size that does not use any excluded vars

arguments

i := index of variable excluded_vars := set of vars that cannot be used in var repr. for an expression, set is usually the union of inactive vars and var i’s representative chain (variables that use var i in its var repr)

return

representation of var i

pub fn substitutions_at_forward(
&mut self,
sm: HashMap<usize, Array1<f32>>,
ordering: Vec<usize>
) -> Option<Vec<(HashMap<usize, Array1<f32>>, Vec<usize>)>>

pub fn cover_vars_by_substitution(
&mut self,
rep_analysis: Vec<(usize, f32)>
) -> HashMap<usize, Array1<f32>>

description

covers vars by substitution by forward procedure:

iterate through cache C of rep_analysis (usually sorted at .1)

pop each var [0]

process its tree structure of vars repr. Re-arrange active vars of vars repr to front of cache C.

arguments

rep_analysis := output from method representative_analysis()

excluded_vars := set of vars that cannot be used in substitution (results in 0’s)

note

will use entire data variable of elements as candidates for var. repr.

pub fn brute_force_cover_vars_by_substitution(
&mut self,
expr: Array1<f32>,
rep_analysis: Vec<(usize, f32)>,
verbose: bool
) -> HashMap<usize, Array1<f32>>

description

uses a cache with elements (substitution map, var search ordering) to search for s-map that produces lowest non-zero unknown active size of expr.

pub fn parent_vars_of_var_in_map(
&mut self,
hm: HashMap<usize, Array1<f32>>,
vi: usize
) -> HashSet<usize>

description

gathers the set of vars that result in var vi as substituent from hm

pub fn representative_decision_smap(
&mut self,
expr: Array1<f32>,
verbose: bool
) -> HashMap<usize, Array1<f32>>

description

outputs a substitution map for expr.

If process_out is on, processes the output expression according to known variables.