Module impliar::enci::implie

struct used to contain element that can generate new elements based on schematic. element. Notation of new elements is in the form of parenthetical notation.

Example: merge n elements

merge s1,t2,u3,v4,w5 results in s1_t2_u3_v4_w5

Example: increment element

increment s1_t2_u3_v4_w5 results in s1_t2_u3_v4_w5(s,t,u,v,w,1)

Example: increment element

increment s1_t2_u3_v4_w5(s,t,u,v,w,1) results in s1_t2_u3_v4_w5(s,t,u,v,w,1)(s,t,u,v,w,2)

Rule for increment

given a multi-chunk, take the last chunk, strip all elements of the chunk of their numbers, merge those elements, and assign them a one.

Calculation of existence and implication

Example: values for merged

existence and implication for s1_t2_u3_v4_w5 is f(s1) * f(t2) * f(u3) * f(v4) * f(w5)

Structs

ImpSetSource

Functions

build_ImpSetSource

merged_elements_string

description

set_source_string_increment

description