Module impliar::enci::seq_encoder

an encoder for an i32 sequence

Structs

FloorDiv

element of IndexFractionNotation

IndexFractionNotation

A sequence encoder that encodes all values of an Array1 by a <Vec> of equal length. Resultant is then able to replicate all indices of the reference sequence and produce an output for any usize.

Functions

build_index_fraction_notation