Struct impliar::metrice::brp::RangePartitionGF2

pub struct RangePartitionGF2 {
    pub f32_vec: Array1<f32>,
    pub binary_labels: Array1<usize>,
    pub size_threshold: usize,
    pub fselect: FSelect,
    pub score: usize,
    pub contra_indices: Vec<usize>,
    fs_cache: Vec<FSelect>,
    c_type: String,
}

A struct that calculates an fs::FSelect instance that fits f32_vec with its corresponding binary_labels (0|1).

The search for this solution can be one of:

greedy approach (see RangePartitionGF2::greedy_search__decision)

brute-force approach (see RangePartitionGF2::brute_force_search__decision)

Fields
f32_vec: Array1<f32>

target vector of f32, elements in [0,1]

binary_labels: Array1<usize>

vector of equal size to <f32_vec>, each i’th element is 0 or 1 and corresponds to i’th element of <f32_vec>

size_threshold: usize

number of partitions that fselect solution can have

fselect: FSelect

solution

score: usize

score of solution

contra_indices: Vec<usize>

indices in f32_vec such that fselect solution contradicts label in binary_labels.

fs_cache: Vec<FSelect>

cache of candidates in brute-force|greedy approach

c_type: String

basic|fm

Implementations

impl RangePartitionGF2

pub fn output(&mut self, f: f32) -> Option<usize>

description

outputs the label of the f32

pub fn is_contra_value(&mut self, f: f32) -> bool

return

if f is a contradicting value according to the contra_indices of fselect.

pub fn best_label_for_bounds(&mut self, b: (f32, f32), r: (usize, usize)) -> usize

return

most frequent label for elements of f32_vec[r.0..r.1] in bounds b

pub fn label_fm(&mut self, f: &mut FSelect, f2: f32)

description

used for mode == “fm”

pub fn modify_and_merge_fselect_bounds(
&mut self,
f: &mut FSelect,
bi: usize,
f2: f32,
r: (usize, usize),
label_mode_majority: bool
) -> ((f32, f32), usize)

description

modify an fselect’s bounds at index bi with float f2

check for intersection with any other bound of fselect and perform appropriate merges.

calculates new label for bound

based on contradiction score

arguments

f := involved FSelect bi := index of bounds of FSelect f f2 := involved f32 r := typically the range 0-(i - 1) right before the i’th element for FSelect f label_mode_majority := bool for relabelling of modified bounds by majority (frequency), alternative is SUM

pub fn brute_force_search__decision(&mut self) -> Option<FSelect>

description

calculates the fs::FSelect with the lowest score by a brute-force search approach

return

fs::FSelect solution or None

pub fn update_cache_fselect_scores(&mut self)

pub fn brute_force_search__decision_at_index(&mut self, i: usize)

description

helper method for brute-force solution, modifies existing fs::FSelect instances in fs_cache by f32_vec[i]`.

pub fn choices_at_index(&mut self, f: FSelect, i: usize) -> Vec<FSelect>

description

outputs the possible f-selects at f32_vec\[i\]

pub fn score_fselect_(
&mut self,
f: &mut FSelect,
r: (usize, usize),
save_contra: bool
) -> f32

description

scores the fs::FSelect f on f32_vec[r.0..r.1] if save_contra is set to true, saves all contradicting indices into struct attribute.

pub fn score_fselect_fm(&mut self, f: &mut FSelect, r: (usize, usize)) -> f32

description

score function based on distance of elements in f32_vec[r.0..r.1] to their labels

pub fn score_fselect(
&mut self,
f: &mut FSelect,
r: (usize, usize),
save_contra: bool
) -> f32

description

calculates an std. score for the fs::FSelect f by observing its performance on the subvector index range r of f32_vec. Formula is: (number of partitions of f) * (number of elements of f32_vec[r.0..r.1] that contradict its binary label)

special case if FSelect violates size threshold t,outputs f32::MAX

pub fn greedy_search__decision(&mut self)

description

greedy search approach to determine the fselect solution

pub fn greedy_search__decision_for_f32(&mut self, i: usize)

description

takes the f32 x at f32_vec[i], and iterates through each fs::FSelect in fs_cache, modifying each one with x, and loads the best modified fs::FSelect into fs_cache.