Struct impliar::metrice::gorillains::GorillaIns

pub struct GorillaIns {

    sequence: Array1<f32>,
    k: usize,
    approach: VRed,
    pub app_out1: Option<f32>,
    pub app_outn: Option<Array1<f32>>,
    wanted_normaln: Option<Array1<usize>>,
    wanted_normal1: Option<usize>,
    pub man_sol: Option<RangePartitionGF2>,
    auto_sol: Option<ArbitraryRangePartition>,
    tail_mode: usize,
    szt: usize,
    pub soln: Option<FSelect>,
    pub corr: Option<Array1<f32>>,
    pub corr2: Option<f32>,

}

Gorilla instructor GorillaIns is a “normal”-detection algorithm that determines if attribute “sequence” (arr1<f32>) is normal.

For a sequence S, applies the approach A, a vreducer::VRed on S by the following A: S –> {0,1}^|S| OR (0|1), based on user arg. (vector of boolean values denoting normal). ….. GorillaIns can proceed by one of the following: pre-labelled data (normal values) for sequence S using data struct RangePartitionGF2 non-labelled data, hypothesis computed by ArbitraryRangePartition ….. GorillaIns can calculate a correction vector (if tail-n) or correction float (if tail-1).

Fields
sequence: Array1<f32>

target of “normal”-analysis

k: usize

number of decimal places considered

approach: VRed

the chained function used to “translate” sequence

app_out1: Option<f32>

tail-1 case output

app_outn: Option<Array1<f32>>

tail-n case output

wanted_normaln: Option<Array1<usize>>

tail-n case wanted normal

wanted_normal1: Option<usize>

tail-1 case wanted normal

man_sol: Option<RangePartitionGF2>

solution for manual approach (provided “wanted”)

auto_sol: Option<ArbitraryRangePartition>

solution for automatic approach (not provided “wanted”)

tail_mode: usize

0 for tai1-1 (output is f32) or 1 for tail-n (output is n-vector)

szt: usize

partition variable

soln: Option<FSelect>

soln from manual or automatic approach

corr: Option<Array1<f32>>

corrector vector for tail-n mode

corr2: Option<f32>

corrector vector for tail-1 mode

Implementations

impl GorillaIns

pub fn predict_sequence(
&mut self,
v: Array1<f32>
) -> (Option<usize>, Option<Array1<usize>>)

description

Predicts the normality of a sequence v by one of the tail approaches, singleton or tail-n (vector).

return: If tail_mode == 0: [0] is solution If tail_mode == 1: [1] is solution

pub fn approach_on_sequence(&mut self) -> (Option<f32>, Option<Array1<f32>>)

description

runs vred::VReducer by tail_mode on sequence

return

tail-1 output or tail-n output

pub fn process_tail1(&mut self) -> (usize, Option<f32>)

description

runs approach_on_sequence and calculates the error-score

return

(label 0|1 of sample,error-cost)

pub fn brute_process_tailn(&mut self)

description

processes tail-n approach by brute-force approach using arp::ArbitraryRangePartition

pub fn improve_approach__labels(
&mut self,
is_multi: bool
) -> (Option<f32>, Option<Array1<f32>>)

description

improves solution by the same size threshold t. uses skewf32::SkewF32 to modify values.

pub fn improve_vred__tailn(&mut self, v: Array1<f32>)

description

improves approach by making a vreducer::VRed adder skew out of v and updating its tail-n w/ it.