Struct impliar::metrice::gorillaj::GorillaJudge

pub struct GorillaJudge {

    pub brx: BatchReader,
    pub bry: Option<BatchReader>,
    pub is_tailn: bool,
    pub data_load: Vec<Array1<f32>>,
    pub label_loadn: Option<Vec<Array1<f32>>>,
    pub label_load1: Option<Vec<f32>>,
    pub base_vr: VRed,
    k: usize,
    reducer_szt: usize,
    bs: usize,
    lbs: usize,
    tail1_skew: Vec<f32>,
    tailn_skew: Vec<SkewF32>,
    bc: GBatchCorrector,
    vr_outputn: Vec<Array1<f32>>,
    vr_output1: Vec<f32>,
    misclass_mtr: f32,
    skew_mtr: f32,

}

gorillaj::GorillaJudge uses gorillains::GorillaIns to learn each sample it processes. Samples are loaded into this data structure by its x-data vcsv::BatchReader and ?y-data? vcsv::BatchReader.

Two metrics are used to gauge the “goodness-of-fit” of the mapping function base_vr: - summation of skew - summation of mis-classification

Each sample s goes through the function base_vr and its skew is the vector I - base_vr(s), I is the vector of f32's in [0,1] such that each i’th value of base_vr(s) should equal the i’th value in I for a labelling of s with no mis-classification error.

Summation of mis-classification is calculated by adding the score of each gorillains::GorillaIns solution (a fs::FSelect instance). The .score attribute of the fs::FSelect instance is its misclassification for the sample.

Fields
brx: BatchReader

x-data reader

bry: Option<BatchReader>

y-data reader (for supervised learning)

is_tailn: bool

is output a vector?

data_load: Vec<Array1<f32>>

batch data from x-data reader

label_loadn: Option<Vec<Array1<f32>>>

batch data from y-data reader (tail-n)

label_load1: Option<Vec<f32>>

batch data from y-data reader (tail-1)

base_vr: VRed

base VReducer used to map each x sample into its y label

k: usize

decimal length for skew values; used by gorillains::GorillaIns

reducer_szt: usize

range-space size for BFGSelectionRule

bs: usize

batch size

lbs: usize

size of last batch read

tail1_skew: Vec<f32>

tail-n case: skew values for each base_vr(s) of s in x-data.

tailn_skew: Vec<SkewF32>

tail-n case: skew values for each base_vr(s) of s in x-data.

bc: GBatchCorrector

batch corrector for refactoring skews

vr_outputn: Vec<Array1<f32>>

tail-n case: output from base_vr to value before y-label

vr_output1: Vec<f32>

tail-1 case: output from base_vr to value before y-label

misclass_mtr: f32

mis-classification score metric

skew_mtr: f32

skew summation score metric

Implementations

impl GorillaJudge

pub fn load_next_batch(&mut self) -> usize

description

loads next batch (x-data, ?y-data?) of size <= bs into memory

return

size of batch

pub fn gorilla_on_batch(&mut self) -> usize

description

pub fn process_gorilla_at_index(
&mut self,
i: usize
) -> (Option<f32>, Option<SkewF32>, Option<f32>, Option<Array1<f32>>, f32)

description

runs a gorillains::GorillaIns instance on the i’th sample in data

return

(f32 skew for tail-1 case, vector skew for tail-n case, vred output-1, vred output-n,misclassification score)

pub fn gorilla_at_index(&mut self, i: usize) -> GorillaIns

pub fn refactor(&mut self) -> Option<f32>

description

return

tail-1 skew if !is_tailn

pub fn refactor_batch_tailn(&mut self) -> (f32, f32, Option<i32>, bool)

description

return

current skew score, skew score after update, skew factor, is adder?

pub fn refactor_batch_tail1(&mut self) -> (f32, f32, f32)

description

uses min,max,mean as points of interest. Refactorization by tail-1 mode only considers adders (no multers).

return:

current score, score after adder, adder

pub fn get_refactor_batch_tail1(&mut self) -> (f32, f32, f32)

description

uses min,max,mean as points of interest. Refactorization by tail-1 mode only considers adders (no multers).

return:

current score, score after adder, adder

pub fn reload_batchcorrctr(&mut self)

description

method called after refactor modification to base_vr.

pub fn update_batchcorrctr(&mut self)

description

pub fn process_next(&mut self, refactor: bool)

description

main function.