Module impliar::metrice::vreducer

Structs

FCast

cast struct for function f, which takes an arr1 as input and outputs another arr1.

FCastF32

cast struct for function f, which takes an arr1 as input and produces another f32 x, and adding ai to x is the output from struct main function .

VRed

structure that acts as a chained-function (a sequence of functions). Each function in the sequence can be of type vreducer::FCast or type skewf32::SkewF32. Main function is .

Functions

build_VRed

sample_fsvecs

sample fvec, svec used for testing.

sample_vred_adder_skew

sample skewf32::SkewF32 used for testing.

std_euclids_reducer

average for gorilla euclid additive vector A and coefficient vector C

std_gcd_reducer