Struct impliar::metrice::vreducer::VRed

pub struct VRed {
    pub fvec: Vec<FCast>,
    pub svec: Vec<SkewF32>,
    pub directions: Vec<usize>,
    pub switch_f: usize,
    pub fi: usize,
    pub si: usize,
    pub tail1: Option<FCastF32>,
    pub tailn: Option<SkewF32>,
}

structure that acts as a chained-function (a sequence of functions). Each function in the sequence can be of type vreducer::FCast or type skewf32::SkewF32. Main function is VRed::apply.

Fields
fvec: Vec<FCast>

sequence of vreducer::FCast.

svec: Vec<SkewF32>

sequence of skewf32::SkewF32.

directions: Vec<usize>

binary sequence with value at each index signifying if fvec function (0) or svec function (1) is used.

switch_f: usize

boolean used to switch between fvec and svec

fi: usize

index of fvec function during vreducer::VRed apply function

si: usize

index of svec function during vreducer::VRed apply function

tail1: Option<FCastF32>

last function in apply, produces a f32 singleton.

tailn: Option<SkewF32>

last function in apply, produces a f32 vector.

Implementations

impl VRed

pub fn apply(
&mut self,
a: Array1<f32>,
tail_type: usize
) -> (Option<f32>, Option<Array1<f32>>)

applies body function (fvec|svec functions) on argument a. If tail_type is 0, applies tail1, o.w. applies tailn.

pub fn apply_body(&mut self, a: Array1<f32>) -> Array1<f32>

pub fn check_directions(&mut self) -> bool

checks validity of directions with respect to function vectors fvec and svec.

pub fn reset_i(&mut self)

index-reset function.

pub fn size_fs(&mut self) -> usize

outputs the number of functions in body function.

pub fn current_switch(&mut self) -> usize

pub fn add_f(&mut self, a: FCast)

adds a vreducer::FCast function to fvec and modifies directions.

pub fn add_s(&mut self, a: SkewF32)

adds a skewf32::SkewF32 to svec and modifies directions.

pub fn mod_tailn(&mut self, nt: SkewF32)

replaces tailn with argument nt.

pub fn mod_tailn_(&mut self, nt: SkewF32)

adds an existing tailn to the end of svec and replaces tailn with nt.

pub fn mod_tail1(&mut self, nt: FCastF32)