Module impliar::setti::bfngsrch

Structs

BFGSearcher

brute-force greedy searcher; designed as a generic structure used to select options given a SelectionRule. Generization is implemented by the function f, external to struct. Function f decides the choice c at column i. The choice c is the input arg to <BFGSearcher.next_srs>. If c is None, searcher takes brute-force approach and considers all available choices at column i. Otherwise, searcher takes choice c at column i

BFGSelectionRule

structure used by brute-force greedy searcher; uses a selection rule each selection rule is accompanied by a binary vector, 1 is greedy search.

Functions

build_BFGSearcher

build_BFGSelectionRule

default_BFGSelectionRule