Module impliar::setti::disinc

file contains a data structure with a name that comes from Disclude-Include, disinc.

Structs

DisIncForwardChainHead

acts as a intersection/difference-like data container

DisIncRule

a struct that uses disinc::Discludia and disinc::Includio to perform pairwise merging of Vec based on float_range

Discludio

difference-space data container

Includia

intersection-space data container

Traits

FAt

Functions

build_DisIncForwardChainHead

hashset_pair_to_incdis_pair

description

set_delta_function_pair_1

description

test_sample_disinc_vec_0

test_sample_disinc_vec_1

test_sample_disinc_vec_2

vec_merge_rule_v1

description