Module impliar::setti::hashfit

Functions

fcollect_max_proper_hashfit_wrt_reference

collects the subset in d with the most matches to array r according to function g.

hashfit_score1

Calculates the fitness of d to reconstruct (fit) r. (number of elements of r activated) - (number of unique elements of d not activated) CAUTION: no arg. check

hashfit_score2

this is the actual functionale used de la fcollect_max_proper_hashfit_wrt_reference.

is_proper_hash_fit

determines if vh is a vector of hashsets that do not contain intersecting elements.

is_proper_hash_fit_arr1

arr1 version of is_proper_hash_fit

next_possible_forward_string_hash_fit

f is the ordered (by d) sub-vector of d. Considers the subvector of d after the last element of f CAUTION: order-check for f not coded