Struct impliar::setti::impli::Impli

pub struct Impli {
    pub kernel: Vec<String>,
    pub impfvec_e: HashMap<String, ImpF>,
    pub impfvec_i: HashMap<String, ImpF>,
    impfvec_fn: Vec<(String, String)>,
    kstatement_fn: (ImpFI32, String),
    options_ratio_fn: (ImpF, String),
    closure_ratio_fn: (ImpF, String),
    ssf: ImpElementSeedSizeF,
    issf: ImpliSSF,
    iss_hash: HashMap<String, ImpSetSource>,
    pub ei_statement: EIStatement,
    current_set_scores: Vec<(Vec<String>, f32)>,
}

Struct used for implication structure impli::Impli.

terminology

Generates a sequence of k-statements from an initial kernel of elements. The sequence is called the ei_statement.

Each k-statement is a sequence of k-vectors (called k-nodes), and k is an arbitrary usize less than or equal to the number of total unique elements. The unique elements of impli::Impli are the kernel elements and the elements generated from qualifying k-nodes in each generated k-statement.

Uses the following generators to aid in generating the next k-statement:

k-value generator (i32): determine the value of k for the next k-statement.

options ratio generator (f32): in range [0,1], determine the total number of unique elements to consider for k-statement.

closure ratio generator (f32): in range [0,1], determine the ratio of total number of combinations of nCr, where n is the number of options and r is the k value.

seed size generator (i32): determines the number of new elements generated from each k-node in the previous k-statement.

algorithm description

for a k-statement, do the following:

fetch options ratio from generator.

rank unique elements of impli::Impli according to the function of implif::ImpliSSf, and fetch those x number of elements according to options ratio.

fetch k-value from generator.

fetch closure ratio from generator.

fetch the appropriate number of combinations according to the nCr and closure ratio values, which constitutes the ei-statement.

update implif::ImpliSSF according to the generated ei-statement.

Fields
kernel: Vec<String>

starting elements

impfvec_e: HashMap<String, ImpF>

kernel new element existence value generator

impfvec_i: HashMap<String, ImpF>

kernel new element implication value generator

impfvec_fn: Vec<(String, String)>

filenames for existence/implication function vectors

kstatement_fn: (ImpFI32, String)

used to determine the k-value of next statement’s vectors

options_ratio_fn: (ImpF, String)

used to determine the options ratio

closure_ratio_fn: (ImpF, String)

used to determine the closure of the options for a k-statement

ssf: ImpElementSeedSizeF

seed size function

issf: ImpliSSF

used to determine the minumum implication of statement

iss_hash: HashMap<String, ImpSetSource>

elements -> their implie::ImpSetSource counter

ei_statement: EIStatement

output of impli::Impli

current_set_scores: Vec<(Vec<String>, f32)>

set scores of last statement sequence of (Vec,f32) pairs.

Implementations

impl Impli

pub fn output_ei_generated(&mut self, k: String, n: usize) -> Vec<(f32, f32)>

description

outputs the first n (existence,implication) values from the generator of element k.

pub fn output_k_generated(&mut self, n: usize) -> Vec<i32>

description

outputs the first n values from the k-value generator.

pub fn output_options_ratio_generated(&mut self, n: usize) -> Vec<f32>

description

outputs the first n values from the options-ratio generator.

pub fn output_closure_ratio_generated(&mut self, n: usize) -> Vec<f32>

description

outputs the first n values from the closure-ratio generator.

pub fn output_statement(&mut self, verbose: bool)

description

calculates the next ei-statement.

pub fn gather_new_elements(&mut self, c: f32, o: Vec<String>, verbose: bool)

description

calculates new elements from k-statement to be loaded, and updates ie_statement.

pub fn add_new_element(&mut self, s: Vec<String>, f: f32, verbose: bool)

description

creates a new implie::ImpSetSource for s and adds s to impf::ImpliSSF

pub fn gather_k_vectors(
&mut self,
o: Vec<String>,
rn: usize,
k: usize,
verbose: bool
)

description

gather all k-vectors (stringized) and their impli::ImpliSSF.f scores.

pub fn strvec_to_impsetsource(&mut self, v: Vec<String>) -> ImpSetSource

description

converts a vector of strings to an implie::ImpSetSource

pub fn gather_options(&mut self, verbose: bool) -> Vec<String>

description

sorted elements by score of implif::ImpliSSF.f, descending order

pub fn ei_of_new_element(&mut self, stridn: String) -> (f32, f32)

description

calculates the (existence,implication) pair of new element with idn stridn

pub fn ei_pair_by_generator_sequence(&mut self, v: Vec<String>) -> (f32, f32)